pyproject.toml

[build-system]
  build-backend = "poetry.core.masonry.api"
  requires = ["poetry-core>=1.0.0"]

[tool]

  [tool.poetry]
    authors = ["Niall Byrne <niall@niallbyrne.ca>"]
    description = "Grocery Shopping App"
    license = "MPL-2"
    name = "panic"
    version = "0.1.0"

    [tool.poetry.dependencies]
      Django = "^3.2.11"
      django-bleach = "^1.0.0"
      django-cors-headers = "^3.9.0"
      django-filter = "^21.1"
      django-split-settings = "^1.0.1"
      django-timezone-field = "^4.2.3"
      django-watchman = "^1.2.0"
      djangorestframework = "^3.12.0"
      djangorestframework-simplejwt = ">=4.7.1,<4.8.0"
      drf-yasg = "^1.20.0"
      gevent = "^21.8.0"
      gunicorn = "^20.1.0"
      langcodes = "^3.1.0"
      pendulum = "2.1.2"
      psycopg2-binary = "^2.9.3"
      python = "^3.7"
      pytz = "^2021.3"

      [tool.poetry.dependencies.Sphinx]
        optional = true
        version = "^4.0.2"

      [tool.poetry.dependencies.dj-rest-auth]
        extras = ["with_social"]
        version = "^1.0.0"

      [tool.poetry.dependencies.sphinxcontrib-django]
        optional = true
        version = "^0.5.1"

    [tool.poetry.dev-dependencies]
      Faker = "^8.11.0"
      bandit = "^1.7.0"
      commitizen = "^2.20.0"
      debugpy = "^1.3.0"
      freezegun = "^1.1.0"
      isort = "^5.8.0"
      mypy = "^0.931"
      parameterized = "^0.8.1"
      pydocstyle = "^6.1.1"
      pylint = "^2.8.3"
      pylint-django = "^2.4.4"
      pytest = "^6.2.5"
      pytest-cov = "^3.0.0"
      pytest-django = "^4.4.0"
      pytest-pylint = "^0.18.0"
      pytest-subtests = "^0.6.0"
      safety = "^1.10.3"
      types-freezegun = "^1.1.0"
      types-pytz = "^2021.3.4"
      yapf = "^0.32.0"
      django-stubs = "^1.9.0"
    djangorestframework-stubs = "^1.4.0"

    [tool.poetry.extras]
      docs = ["Sphinx", "sphinxcontrib-django"]
